requirements.txt
# ============================================================
# AI PM Autonomous Learning & Intelligence System
# requirements.txt — TDD v2.0 §Dependencies
# ============================================================

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication & Security
fastapi-csrf-protect==0.3.2
slowapi==0.1.9
python-multipart==0.0.6  # Required for form data parsing

# Google APIs
google-generativeai==0.3.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Content Extraction — Layered chain (TDD v2.0 §Utils)
feedparser==6.0.10
trafilatura==1.6.3
readability-lxml==0.8.1
newspaper3k==0.2.8
chardet==5.2.0
lxml==4.9.4
lxml_html_clean==0.1.0

# Text Processing — Fuzzy dedup (FRD FS-04.2)
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# HTTP Clients
httpx==0.26.0
requests==2.31.0
tenacity==8.2.3

# Templates & Charts
jinja2==3.1.3
matplotlib==3.8.2

# Logging — loguru structured JSON (TDD §Logging)
loguru==0.7.2

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Crypto / Hashing
cryptography==42.0.0

# Testing (dev — not deployed)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0
pytest-cov==4.1.0
responses==0.24.1
httpx==0.26.0
